BUGFIX for postgress db type

+++ ext_tables.sql
@@ -5,7 +5,7 @@ CREATE TABLE tx_weather2_domain_model_currentweather (
 	name varchar(255) DEFAULT '' NOT NULL,
 	measure_timestamp int(11) DEFAULT '0' NOT NULL,
 	temperature_c double(4,2) DEFAULT '0.0' NOT NULL,
-	pressure_hpa int(11) DEFAULT '0' NOT NULL,
+	pressure_hpa double(4,2) DEFAULT '0' NOT NULL,
 	humidity_percentage int(11) unsigned DEFAULT '0' NOT NULL,
 	min_temp_c double(4,2) DEFAULT '0.0' NOT NULL,
 	max_temp_c double(4,2) DEFAULT '0.0' NOT NULL,
